separate basic running avg from premium
