+++ components/dashboard/metrics.tsx
@@ -6,11 +6,10 @@ import { useQuery } from "@apollo/client"
 import { DashboardFilterContext } from "./context"
 
 export const Metrics: React.FC = () => {
-  const { runningAvg, setDaysOfUse, setAvgHours } = useContext(
-    DashboardFilterContext
-  )
+  const { premium, basicRunningAvg, premiumRunningAvg, setDaysOfUse, setAvgHours } =
+    useContext(DashboardFilterContext)
   const { data, loading } = useQuery(DASBOARD_METRICS_QUERY, {
-    variables: { runningAvg },
+    variables: { runningAvg: premium ? premiumRunningAvg[0] : basicRunningAvg[0] },
   })
 
   const dashboardMetrics = data?.dashboardMetrics.dashboardMetrics
