+++ types/dashboard.ts
@@ -11,7 +11,7 @@ interface PgDashboardMetrics {
 }
 
 type DashboardFilters = {
-  runningAvg: RunningAverage | null
+  premiumRunningAvg: RunningAverage | null
   analyzeEntities: string | null
   tokenTags: string | null
   minTokenCount: string | null
@@ -20,6 +20,7 @@ type DashboardFilters = {
   hiddenTokens: string | null
   hiddenEntities: string | null
   sentencesRating: string | null
+  basicRunningAvg: RunningAverage | null
   minWordCount: string | null
   basicSentencesRating: string | null
   hiddenWords: string | null
